requirements.txt
psycopg2-binary
python-dotenv
telethon
pillow
pytesseract
